Refactor Solr client to use HttpJdkSolrClient and update utils
Replaces usage of CloseableHttpClient and HttpSolrClient with HttpJdkSolrClient in Solr integration, removes TurSolrConfiguration, and updates related classes and constructors. Renames TurSNSiteFieldUtils to TurSNSiteFieldService and moves it to a new package. Refactors TurUtilTests to TurUtils as a utility class and updates test imports accordingly. Also removes unnecessary exception declaration in TurSecurityConfigProduction.