Refactor ORDER enum to Order in TurSNQuery and update usages
Renamed the ORDER enum to Order in TurSNQuery for improved Java naming conventions. Updated all references in source, test, and sample code, as well as related classes such as TurSNSortField and TurToolsService. Also added central-publishing-maven-plugin to commons and sdk pom.xml files, and moved plugin configuration to pluginManagement in the root pom.xml.